+++ src/requests/account/blocklist/get_blocked_accounts.rs
@@ -22,12 +22,12 @@ impl TryFrom<Response> for Vec<Account> {
 pub(crate) struct GetBlockedAccountsRequest {
     #[serde(rename = "accountId")]
     id: u64,
-    offset: u64,
+    offset: usize,
 }
 impl GetBlockedAccountsRequest {
     pub(crate) const PAGE_SIZE: usize = 20;
 
-    pub(crate) fn new(id: u64, offset: u64) -> Self {
+    pub(crate) fn new(id: u64, offset: usize) -> Self {
         Self { id, offset }
     }
 }
